given an error string, print out some context
New exported functions `explain` and `wtf`.

These should be called like `explain("some error message")`
and will display to the user some info about the likely
cause of an error message.

At present, the only error description that is available
is for non-subsettability. An apology function is also
present which apologises if we can't explain some error
message.